# docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.MetadataResponse:
    properties:
      pagination:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.PaginationResponse'
      request_id:
        example: req-12345
        type: string
      timestamp:
        example: "2023-10-27T10:30:00Z"
        type: string
    type: object
...
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.SuccessResponse:
    properties:
      data:
        description: Use omitempty if data might be nil/empty
      message:
        description: Optional success message
        type: string
      meta:
        $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.MetadataResponse'
      success:
        example: true
        type: boolean
    type: object
...
  github_com_threatflux_dockerServerMangerGoMCP_internal_models.UserListResponse:
    description: Contains a list of user details along with pagination information.
    properties:
      metadata:
        allOf:
        - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.MetadataResponse'
        description: |-
          Metadata contains pagination and other metadata for the response.
          required: true
      users:
        description: |-
          Users is the list of user objects returned for the current page.
          required: true
        items:
          $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.UserResponse'
        type: array
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the API documentation for the Docker Server Manager Go MCP.
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: Docker Server Manager Go MCP API
  version: "1.0"
paths:
  /admin/users:
    get:
      description: Retrieves a paginated list of all users. Only admins can perform
        this action.
      parameters:
      - default: 1
        description: Page number
        in: query
        name: page
        type: integer
      - default: 10
        description: Number of items per page
        in: query
        maximum: 100
        minimum: 1
        name: page_size
        type: integer
      - default: id
        description: Field to sort by (e.g., email, name, created_at)
        in: query
        name: sort_by
        type: string
      - default: asc
        description: Sort order (asc, desc)
        enum:
        - asc
        - desc
        in: query
        name: sort_order
        type: string
      - description: Search term for email or name
        in: query
        name: search
        type: string
      - description: Filter by active status
        in: query
        name: active
        type: boolean
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved users" // Use UserListResponse
          schema:
            allOf:
            - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.UserListResponse'
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "403":
          description: Admin privileges required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List all users (Admin)
      tags:
      - Admin
    post:
      consumes:
      - application/json
      description: Creates a new user with specified details. Only admins can perform
        this action.
      parameters:
      - description: New user details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.AdminCreateUserRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully created user
          schema:
            allOf:
            - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.UserResponse'
              type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "403":
          description: Admin privileges required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "409":
          description: Email already in use
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create a new user (Admin)
      tags:
      - Admin
  /admin/users/{id}:
    delete:
      description: Deletes a user by ID. Only admins can perform this action.
      parameters:
      - description: User ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "204":
          description: Successfully deleted user
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "403":
          description: Admin privileges required or cannot delete self
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete a user (Admin)
      tags:
      - Admin
    get:
      description: Retrieves details for a specific user by their ID. Only admins
        can perform this action.
      parameters:
      - description: User ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved user details
          schema:
            allOf:
            - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.UserResponse'
              type: object
        "400":
          description: Invalid user ID
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "403":
          description: Admin privileges required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user by ID (Admin)
      tags:
      - Admin
    put:
      consumes:
      - application/json
      description: Updates a user's details by ID. Only admins can perform this action.
      parameters:
      - description: User ID
        example: 1
        in: path
        name: id
        required: true
        type: integer
      - description: User update details (fields are optional)
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.AdminUpdateUserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully updated user
          schema:
            allOf:
            - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.UserResponse'
              type: object
        "400":
          description: Invalid input or user ID
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "403":
          description: Admin privileges required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "404":
          description: User not found
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "409":
          description: Email already in use
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update a user (Admin)
      tags:
      - Admin
  /api/v1/containers/{id}/exec:
    post:
      consumes:
      - application/json
      description: |-
        Creates and starts an exec instance in a running container.
        Supports both non-interactive execution (HTTP POST) and interactive sessions (WebSocket upgrade).
        For interactive sessions, upgrade the connection to WebSocket.
      parameters:
      - description: Container ID
        example: my-running-container
        in: path
        name: id
        required: true
        type: string
      - description: Exec configuration
        in: body
        name: exec
        required: true
        schema:
          $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ContainerExecCreateRequest'
      produces:
      - application/json
      responses:
        "101":
          description: Switching Protocols (interactive WebSocket)
          schema:
            type: string
        "200":
          description: Success (non-interactive)
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ContainerExecResponse'
        "400":
          description: Bad Request (e.g., missing command, invalid ID)
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "403":
          description: Forbidden (user lacks permission)
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "404":
          description: Container Not Found
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Execute Command in Container
      tags:
      - Containers
...
  /auth/login:
    post:
      consumes:
      - application/json
      description: Logs in a user with email and password, returning JWT tokens.
      parameters:
      - description: User login credentials
        in: body
        name: credentials
        required: true
        schema:
          $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully logged in
          schema:
            allOf:
            - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.TokenResponse'
              type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "401":
          description: Invalid credentials
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      summary: Log in a user
      tags:
      - Auth
  /auth/logout:
    post:
      description: Invalidates the current access token.
      responses:
        "204":
          description: Successfully logged out
        "400":
          description: Invalid authorization header
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Log out a user
      tags:
      - Auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Refreshes the access and refresh tokens using a valid refresh token.
      parameters:
      - description: Refresh token details
        in: body
        name: token
        required: true
        schema:
          $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.RefreshTokenRequest'
      produces:
      - application/json
      responses:
        "200":
          description: Successfully refreshed tokens
          schema:
            allOf:
            - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.TokenResponse'
              type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "401":
          description: Invalid or expired refresh token
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      summary: Refresh JWT tokens
      tags:
      - Auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Registers a new user. The first user registered automatically gets
        the admin role.
      parameters:
      - description: User registration details
        in: body
        name: user
        required: true
        schema:
          $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Successfully registered and logged in
          schema:
            allOf:
            - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.TokenResponse'
              type: object
        "400":
          description: Invalid input
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "409":
          description: Email already in use
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      summary: Register a new user
      tags:
      - Auth
  /compose:
    get:
      description: Get a list of tracked Docker Compose deployments and their status.
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved deployments
          schema:
            allOf:
            - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.DeploymentInfo'
                  type: array
              type: object
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: List Compose Deployments
      tags:
      - Compose
  /compose/{id}:
    get:
      description: Get detailed status information for a specific Docker Compose deployment
        by project name.
      parameters:
      - description: Project Name (ID)
        example: my-web-app
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Successfully retrieved deployment details
          schema:
            allOf:
            - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.SuccessResponse'
            - properties:
                data:
                  $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.DeploymentInfo'
              type: object
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "404":
          description: Deployment not found
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get Compose Deployment Details
      tags:
      - Compose
  /compose/{id}/down:
    post:
      description: Stops and removes containers, networks, and optionally volumes
        for a deployment.
      parameters:
      - description: Project Name (ID)
        example: my-web-app
        in: path
        name: id
        required: true
        type: string
      - default: false
        description: Remove named volumes declared in the 'volumes' section
        example: true
        in: query
        name: remove_volumes
        type: boolean
      - default: false
        description: Remove containers for services not defined in the Compose file
        example: true
        in: query
        name: remove_orphans
        type: boolean
      - default: false
        description: Force removal
        example: false
        in: query
        name: force
        type: boolean
      produces:
      - application/json
      responses:
        "202":
          description: Removal process started
          schema:
            allOf:
            - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.SuccessResponse'
            - properties:
                message:
                  type: string
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "404":
          description: Deployment not found (required for ComposeFile)
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Remove Compose Project (Down)
      tags:
      - Compose
  /compose/{id}/restart:
    post:
      description: Restarts containers for services in a deployment.
      parameters:
      - description: Project Name (ID)
        example: my-web-app
        in: path
        name: id
        required: true
        type: string
      - description: Timeout in seconds for stopping containers before restarting
        example: 5
        in: query
        name: timeout
        type: integer
      produces:
      - application/json
      responses:
        "202":
          description: Restart process started
          schema:
            allOf:
            - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.SuccessResponse'
            - properties:
                message:
                  type: string
              type: object
        "400":
          description: Invalid query parameters
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "404":
          description: Deployment not found (required for ComposeFile)
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Restart Compose Project Services
      tags:
      - Compose
  /compose/{id}/scale:
    post:
      consumes:
      - application/json
      description: Scales a specific service within a deployment to the desired number
        of replicas.
      parameters:
      - description: Project Name (ID)
        example: my-web-app
        in: path
        name: id
        required: true
        type: string
      - description: Service name and replica count
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ComposeScaleRequest'
      produces:
      - application/json
      responses:
        "202":
          description: Scaling process started
          schema:
            allOf:
            - $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.SuccessResponse'
            - properties:
                message:
                  type: string
              type: object
        "400":
          description: Invalid request body or parameters
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "401":
          description: Authentication required
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "404":
          description: Deployment or service not found
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
        "500":
          description: Internal server error
          schema:
            $ref: '#/definitions/github_com_threatflux_dockerServerMangerGoMCP_internal_models.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Scale Compose Service
      tags:
      - Compose
...
schemes:
- http
- https
securityDefinitions:
  BearerAuth:
    description: 'Type "Bearer" followed by a space and JWT token. Example: "Bearer
      {token}"'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"